Fixed Repository and added Specs
Fixed the SesameProtocol module to add the size, context, and namespaces
methods when used as extension.
Fixed the TransactionBuilder to add datatypes and create the whole
transaction xml document.
Added specs to add and remove triples from the test triples store.  Also
added before(:all) and after(:all) to populate and clear the triple
store for the ActiveSesame spec description.